refactor(car): 랜덤 로직 분리 및 매직넘버 상수화
