+++ README.md
@@ -89,5 +89,5 @@
 - [X] 사용자로부터 이동을 시도할 횟수, 자동차 수를 받아라
   - [X] 입력 가이드를 출력하라
 - [X] 자동차 수만큼 자동차를 생성하라
-- [ ] 시도 횟수만큼 자동차를 움직여라
+- [X] 시도 횟수만큼 자동차를 움직여라
 - [ ] 자동차의 상태를 출력하라
