+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import racingcar.domain.Car;
 import racingcar.ui.InputView;
 
 import java.util.ArrayList;
