+++ src/main/java/racingcar/domain/RandomGenerator.java
@@ -1,0 +1,14 @@
+package racingcar.domain;
+
+import java.util.Random;
+
+public class RandomGenerator {
+
+    private static final Random RANDOM = new Random();
+    private static final int BOUND = 10;
+
+    public static int getRandomNumber() {
+        return RANDOM.nextInt(BOUND);
+    }
+
+}
